+++ cnxpublishing/subscribers.py
@@ -6,7 +6,7 @@ from pyramid.events import subscriber
 from pyramid.threadlocal import get_current_registry
 
 
-from . import events
+from . import events, utils
 from .bake import remove_baked, bake
 from .db import (
     update_module_state,
@@ -30,7 +30,21 @@ def track_baking_proc_state(result, module_ident, cursor):
 def post_publication_processing(event, cursor):
     """Process post-publication events coming out of the database."""
     module_ident, ident_hash = event.module_ident, event.ident_hash
-    logger.debug('Processing module_ident={} ident_hash={}'.format(
+
+    celery_app = get_current_registry().celery_app
+
+    # Check baking is not already queued.
+    cursor.execute('SELECT result_id::text '
+                   'FROM document_baking_result_associations '
+                   'WHERE module_ident = %s', (module_ident,))
+    for result in cursor.fetchall():
+        state = celery_app.AsyncResult(result[0]).state
+        if state in ('QUEUED', 'STARTED', 'RETRY'):
+            logger.debug('Already queued module_ident={} ident_hash={}'.format(
+                module_ident, ident_hash))
+            return
+
+    logger.debug('Queued for processing module_ident={} ident_hash={}'.format(
         module_ident, ident_hash))
     update_module_state(cursor, module_ident, 'processing', None)
     # Commit the state change before preceding.
@@ -39,8 +53,9 @@ def post_publication_processing(event, cursor):
     # Start of task
     # FIXME Looking up the task isn't the most clear usage here.
     task_name = 'cnxpublishing.subscribers.baking_processor'
-    baking_processor = get_current_registry().celery_app.tasks[task_name]
+    baking_processor = celery_app.tasks[task_name]
     result = baking_processor.delay(module_ident, ident_hash)
+    baking_processor.backend.store_result(result.id, None, 'QUEUED')
 
     # Save the mapping between a celery task and this event.
     track_baking_proc_state(result, module_ident, cursor)
@@ -74,9 +89,46 @@ def _get_recipe_ids(module_ident, cursor):
     return cursor.fetchone()
 
 
-@task()
+@task(bind=True)
 @with_db_cursor
-def baking_processor(module_ident, ident_hash, cursor=None):
+def baking_processor(self, module_ident, ident_hash, cursor=None):
+    if self.request.retries == 0:
+        cursor.execute("""\
+SELECT module_ident, ident_hash(uuid, major_version, minor_version)
+FROM modules NATURAL JOIN modulestates
+WHERE uuid = %s AND statename IN ('post-publication', 'processing')
+ORDER BY major_version DESC, minor_version DESC""",
+                       (utils.split_ident_hash(ident_hash)[0],))
+        latest_module_ident = cursor.fetchone()
+        if latest_module_ident:
+            if latest_module_ident[0] != module_ident:
+                logger.debug("""\
+More recent version (module_ident={} ident_hash={}) in queue. \
+Move this message (module_ident={} ident_hash={}) \
+to the deferred (low priority) queue"""
+                             .format(latest_module_ident[0],
+                                     latest_module_ident[1],
+                                     module_ident, ident_hash))
+
+                raise self.retry(queue='deferred')
+        else:
+            # In case we can't find the latest version being baked, we'll
+            # continue with baking this one
+            pass
+
+    logger.debug('Starting baking module_ident={} ident_hash={}'
+                 .format(module_ident, ident_hash))
+
+    recipe_ids = _get_recipe_ids(module_ident, cursor)
+
+    state = 'current'
+    if recipe_ids == (None, None):
+        remove_baked(ident_hash, cursor=cursor)
+        logger.debug('Finished unbaking module_ident={} ident_hash={} '
+                     'with a final state of \'{}\'.'
+                     .format(module_ident, ident_hash, state))
+        update_module_state(cursor, module_ident, state, None)
+        return
 
     try:
         binder = export_epub.factory(ident_hash)
@@ -88,7 +140,7 @@ def baking_processor(module_ident, ident_hash, cursor=None):
         update_module_state(cursor, module_ident, 'errored', None)
         raise
     finally:
-        logger.debug('Finished exporting module_ident={} ident_hash={}'
+        logger.debug('Exported module_ident={} ident_hash={}'
                      .format(module_ident, ident_hash))
 
     cursor.execute("""\
@@ -98,9 +150,6 @@ WHERE ident_hash(uuid, major_version, minor_version) = %s""",
     publisher, message = cursor.fetchone()
     remove_baked(ident_hash, cursor=cursor)
 
-    recipe_ids = _get_recipe_ids(module_ident, cursor)
-
-    state = 'current'
     for recipe_id in recipe_ids:
         try:
             bake(binder, recipe_id, publisher, message, cursor=cursor)
@@ -116,7 +165,7 @@ WHERE ident_hash(uuid, major_version, minor_version) = %s""",
                 update_module_state(cursor, module_ident, state, recipe_id)
                 raise
         finally:
-            logger.debug('Finished module_ident={} ident_hash={} '
+            logger.debug('Finished baking module_ident={} ident_hash={} '
                          'with a final state of \'{}\'.'
                          .format(module_ident, ident_hash, state))
             update_module_state(cursor, module_ident, state, recipe_id)