+++ docs/changes.rst
@@ -1,9 +1,49 @@
+
 .. Use the following to start a new version entry:
 
    |version|
    ----------------------
 
    - feature message
+
+0.9.3
+-----
+
+- Explicitly close all psycopg2 db connections (#187)
+- Refactor and fix content-status view (#186)
+
+0.9.2
+-----
+
+- Check for a traceback when handling a celery task failure (#185)
+
+0.9.1
+-----
+
+- Make sure to reserve uuids for new composite content (#184)
+
+0.9.0
+-----
+
+- Use default icon for unknown states on content-status page (#182)
+- Fix to not error when no recipe is found (#180)
+- Optimize post publishing queue (#175)
+- Reword baking procedure log messages (#174)
+- Fix to add view templates to the package distribution (#169)
+- Allow content status pages to be publicly visible (#171)
+- Add views to view and inspect the content publication status (#161)
+- Add a workaround an issue with celery tests, which allows us
+  to unskip them (#170)
+- Fix tests by adding an empty ruleset file
+- Fix tests for change in bake() function signature
+- Fix to fetch recipe text durning baking
+- Use print-style to select recipe and fallback (#162)
+- Add admin page for managing site banner messages (#163)
+
+0.8.1
+-----
+
+- Check for a traceback when handling a celery task failure (#185)
 
 0.8.0
 -----